Race works pretty much completey again, wooh!
