+++ SimTECH/SimTECH/PageModels/RaceWeekModel.cs
@@ -1,5 +1,4 @@
 ﻿using SimTECH.Data.Models;
-using SimTECH.PageModels.Racing;
 
 namespace SimTECH.PageModels
 {
@@ -14,123 +13,5 @@ namespace SimTECH.PageModels
         public List<RaweCeekDriver> RaweCeekDrivers { get; set; }
 
         public List<long> ConsumablePenalties { get; set; } = new();
-    }
-
-    public class RaweCeekDriver
-    {
-        public long SeasonDriverId { get; set; }
-        public string FirstName { get; set; } = string.Empty;
-        public string LastName { get; set; } = string.Empty;
-        public Country Nationality { get; set; }
-        public int Number { get; set; }
-        public TeamRole Role { get; set; }
-
-        public long SeasonTeamId { get; set; }
-        public string TeamName { get; set; }
-        public string Colour { get; set; }
-        public string Accent { get; set; }
-
-        public long ManufacturerId { get; set; }
-        public string ManufacturerName { get; set; }
-        public string ManufacturerColour { get; set; }
-        public string ManufacturerAccent { get; set; }
-
-        public int QualyPower { get; set; }
-        public int RacePower { get; set; }
-
-        public int Attack { get; set; }
-        public int Defense { get; set; }
-        public int DriverReliability { get; set; }
-        public int CarReliability { get; set; }
-        public int EngineReliability { get; set; }
-        public int WearMinMod { get; set; }
-        public int WearMaxMod { get; set; }
-        public int RngMinMod { get; set; }
-        public int RngMaxMod { get; set; }
-        public int LifeBonus { get; set; }
-        public bool HasFastestLap { get; set; }
-
-        public long? PenaltyId { get; set; }
-        public int? Penalty { get; set; }
-        public string? Reasons { get; set; }
-
-        public long ResultId { get; set; }
-        public int Grid { get; set; }
-        public int Position { get; set; }
-        public int Score { get; set; }
-        public RaceStatus Status { get; set; }
-        //public int Setup { get; set; }
-        public int TyreLife { get; set; }
-        public bool FastestLap { get; set; }
-        public int Overtaken { get; set; }
-        public int Defended { get; set; }
-        public long? IncidentId { get; set; }
-        public Tyre Tyre { get; set; }
-
-        public string FullName => FirstName + " " + LastName;
-    }
-
-    public static class ExtendRaweCeekDriver
-    {
-        public static SessionDriver MapToSessionDriver(this RaweCeekDriver driver, int amountRuns)
-        {
-            return new SessionDriver
-            {
-                SeasonDriverId = driver.SeasonDriverId,
-                ResultId = driver.ResultId,
-                FirstName = driver.FirstName,
-                LastName = driver.LastName,
-                Nationality = driver.Nationality,
-                Number = driver.Number,
-                TeamName = driver.TeamName,
-                Colour = driver.Colour,
-                Accent = driver.Accent,
-                Power = driver.QualyPower,
-                Scores = new int[amountRuns]
-            };
-        }
-
-        public static RaceDriver MapToRaceDriver(this RaweCeekDriver driver)
-        {
-            return new RaceDriver
-            {
-                ResultId = driver.ResultId,
-                SeasonDriverId = driver.SeasonDriverId,
-                SeasonTeamId = driver.SeasonTeamId,
-                FirstName = driver.FirstName,
-                LastName = driver.LastName,
-                FullName = driver.FullName,
-                Nationality = driver.Nationality,
-                Number = driver.Number,
-                Role = driver.Role,
-
-                TeamName = driver.TeamName,
-                Colour = driver.Colour,
-                Accent = driver.Accent,
-
-                Power = driver.RacePower,
-                Attack = driver.Attack,
-                Defense = driver.Defense,
-
-                Status = driver.Status,
-                Position = driver.Position,
-                Grid = driver.Grid,
-                //Setup = driver.Setup,
-                TyreLife = driver.TyreLife,
-                CurrentTyre = driver.Tyre,
-                DriverReliability = driver.DriverReliability,
-                CarReliability = driver.CarReliability,
-                EngineReliability = driver.EngineReliability,
-                WearMinMod = driver.WearMinMod,
-                WearMaxMod = driver.WearMaxMod,
-                RngMinMod = driver.RngMinMod,
-                RngMaxMod = driver.RngMaxMod,
-                LifeBonus = driver.LifeBonus,
-
-                HasFastestLap = driver.HasFastestLap,
-                OvertakeCount = driver.Overtaken,
-                DefensiveCount = driver.Defended,
-            };
-        }
     }
 }
