+++ SimTECH/SimTECH/Pages/Index.razor
@@ -205,6 +205,6 @@
     private string DisplayLeagueName(long leagueId) => activeLeagues.Find(e => e.Id == leagueId)?.Name ?? "[Unknown]";
 
     private void GoToSeason(long seasonId) => _nav.NavigateTo($"/overview/{seasonId}");
-    private void GoToPostRace(long raceId) => _nav.NavigateTo($"/postrace/{raceId}");
+    private void GoToPostRace(long raceId) => _nav.NavigateTo($"/raceweek/{raceId}");
     private void GoToRaceWeek(long raceId) => _nav.NavigateTo($"/raceweek/{raceId}");
 }
