+++ SimTECH/SimTECH/Pages/Racing/Practice.razor
@@ -1,31 +1,29 @@
-﻿<MudContainer Class="mt-4">
-    @if (!loading)
-    {
-        <MudPaper MinHeight="60px">
-            <MudStack Row="true" AlignItems="AlignItems.Center" Justify="Justify.FlexEnd">
-                <MudButtonGroup Color="Color.Tertiary" Variant="Variant.Outlined">
-                    @if (!tempIsFinished)
+﻿@if (!loading)
+{
+    <MudPaper MinHeight="60px">
+        <MudStack Row="true" AlignItems="AlignItems.Center" Justify="Justify.FlexEnd">
+            <MudButtonGroup Color="Color.Tertiary" Variant="Variant.Outlined">
+                @if (!isFinished)
+                {
+                    @if (advancedRuns == amountRuns)
                     {
-                        @if (advancedRuns == amountRuns)
-                        {
-                            <MudButton OnClick="Finish">Finish</MudButton>
-                        }
-                        else
-                        {
-                            <MudButton OnClick="Advance">Advance</MudButton>
-                        }
+                        <MudButton OnClick="Finish">Finish</MudButton>
                     }
-                </MudButtonGroup>
-            </MudStack>
-        </MudPaper>
-        <MudPaper>
-            <CommonSession Drivers="@practiceDrivers"
-                           Title="@raceName"
-                           Country="@raceCountry"
-                           AmountRuns="@amountRuns" />
-        </MudPaper>
-    }
-</MudContainer>
+                    else
+                    {
+                        <MudButton OnClick="Advance">Advance</MudButton>
+                    }
+                }
+            </MudButtonGroup>
+        </MudStack>
+    </MudPaper>
+    <MudPaper>
+        <CommonSession Drivers="@practiceDrivers"
+                       Title="@raceName"
+                       Country="@raceCountry"
+                       AmountRuns="@amountRuns" />
+    </MudPaper>
+}
 
 @code {
     [CascadingParameter] public RaweCeekModel RaweCeek { get; set; }
@@ -38,7 +36,7 @@
     private List<SessionDriver> practiceDrivers = new();
 
     bool loading = true;
-    bool tempIsFinished = false;
+    bool isFinished = false;
     string raceName = string.Empty;
     Country raceCountry = Constants.DefaultCountry;
     int amountRuns;
@@ -51,7 +49,7 @@
     protected override void OnInitialized()
     {
         gapMarge = Config.GapMarge;
-        tempIsFinished = !OnFinish.HasDelegate;
+        isFinished = !OnFinish.HasDelegate;
     }
 
     protected override void OnParametersSet()
@@ -67,8 +65,9 @@
 
         var currentSessionScores = PracticeScores.Where(e => e.Index == PracticeIndex).ToList();
 
+        // Should be refactored when saving results on click instead of on finish
         if (currentSessionScores.Any())
-            advancedRuns = currentSessionScores.Select(e => e.Index).Max();
+            isFinished = true;
 
         foreach (var driver in RaweCeek.RaweCeekDrivers)
         {
@@ -120,6 +119,6 @@
         PracticeScores.AddRange(newScores);
 
         await OnFinish.InvokeAsync(PracticeIndex);
-        tempIsFinished = true;
+        isFinished = true;
     }
 }
